+++ .circleci/config.yml
@@ -66,18 +66,18 @@ jobs:
           command: npm ci --prefer-offline --ignore-scripts --no-audit --no-fund
       - run:
           name: Run tests
-          command: npm run test:ci
+          command: npm run test:ci -w email-misspelled
       - run:
           name: Run test coverage
-          command: npm run test:coverage
+          command: npm run test:coverage -w email-misspelled
       - store_artifacts:
           path: tests/coverage/lcov-report
       - run:
           name: Report coverage
-          command: npm run test:coverage:report
+          command: npm run test:coverage:report -w email-misspelled
       - run:
           name: Run build
-          command: npm run build
+          command: npm run build -w email-misspelled
       - save_cache:
           paths:
             - /node_modules
@@ -94,7 +94,7 @@ jobs:
       - *print_information
       - run:
           name: Publish package
-          command: npm publish --access public --tag alpha
+          command: npm publish --access public --tag alpha -w email-misspelled
       - *persist
 
   publish:
@@ -107,7 +107,7 @@ jobs:
       - *print_information
       - run:
           name: Publish package
-          command: npm publish --access public
+          command: npm publish --access public -w email-misspelled
       - *persist
 
 workflows:
