+++ packages/email-misspelled/.eslintrc
