+++ packages/email-misspelled/.npmignore
