+++ packages/email-misspelled/commitlint.config.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line
+export default { extends: ["@commitlint/config-conventional"] }
