+++ packages/email-misspelled/src/domains/README.md
