+++ packages/email-misspelled/src/domains/hotmail.ts
