+++ packages/email-misspelled/src/domains/index.ts
