+++ packages/email-misspelled/src/domains/microsoft.ts
