+++ packages/email-misspelled/src/domains/outlook.ts
