+++ packages/email-misspelled/src/domains/yahoo.ts
