+++ packages/email-misspelled/src/helpers/corrector.ts
