+++ packages/email-misspelled/src/helpers/domain-mapper.ts
