+++ packages/email-misspelled/src/helpers/get-domain.ts
