+++ packages/email-misspelled/src/index.ts
