+++ packages/email-misspelled/tsconfig.cjs.json
