+++ packages/email-misspelled/tsconfig.json
